refactor(controllers): extract HelpVersionGate from HelpSessionController
- Move probeAssistantVersion, checkAssistantVersion, and the manual
  Check-again cooldown state/logic into a sibling module
- Controller delegates checkVersionAgain() to a HelpVersionGate instance
  wired through a narrow host interface (getSnapshot/patch/lastInputs/
  maybeAutoLaunch) — no behavior change
